# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FoodMart API",
        "description": "FoodMart E-commerce API Documentation",
        "contact": {
            "name": "FoodMart Support",
            "email": "karim.hossam813@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/search": {
            "get": {
                "tags": [
                    "Search"
                ],
                "summary": "Search products",
                "description": "Search for products with various filters like keyword, category, price range, and sorting",
                "operationId": "searchProducts",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Search keyword",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "apple"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Category slug",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "fruits-vegetables"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Product type",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "fruits"
                        }
                    },
                    {
                        "name": "min_price",
                        "in": "query",
                        "description": "Minimum price",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "example": 5
                        }
                    },
                    {
                        "name": "max_price",
                        "in": "query",
                        "description": "Maximum price",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "example": 50
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort option",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "latest",
                                "price_asc",
                                "price_desc",
                                "name_asc",
                                "name_desc"
                            ],
                            "example": "latest"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Results per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 12
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "5 products found"
                                        },
                                        "data": {
                                            "properties": {
                                                "query": {
                                                    "type": "string",
                                                    "example": "apple"
                                                },
                                                "filters": {
                                                    "type": "object"
                                                },
                                                "products": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Fresh Apples"
                                                            },
                                                            "slug": {
                                                                "type": "string",
                                                                "example": "fresh-apples"
                                                            },
                                                            "price": {
                                                                "type": "number",
                                                                "example": 4.99
                                                            },
                                                            "formatted_price": {
                                                                "type": "string",
                                                                "example": "$4.99"
                                                            },
                                                            "in_stock": {
                                                                "type": "boolean",
                                                                "example": true
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "pagination": {
                                                    "properties": {
                                                        "current_page": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "last_page": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "per_page": {
                                                            "type": "integer",
                                                            "example": 12
                                                        },
                                                        "total": {
                                                            "type": "integer",
                                                            "example": 5
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/search/quick": {
            "get": {
                "tags": [
                    "Search"
                ],
                "summary": "Quick search (autocomplete)",
                "description": "Get quick search results for autocomplete functionality. Returns max 8 results.",
                "operationId": "quickSearchProducts",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Search keyword (minimum 2 characters)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "app"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Fresh Apples"
                                                    },
                                                    "price": {
                                                        "type": "string",
                                                        "example": "$4.99"
                                                    },
                                                    "type": {
                                                        "type": "string",
                                                        "example": "Fruits"
                                                    },
                                                    "image": {
                                                        "type": "string",
                                                        "example": "http://localhost:8000/storage/products/apple.jpg"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your Bearer token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}